refactor(render): extract spinner (makeSpinner/spinnerPhase) to render/spinner.ts
